Time: 0 ms (100.00%) | Memory: 11.7 MB (99.42%) - LeetSync

+++ 797-rabbits-in-forest/rabbits-in-forest.cpp
@@ -1,6 +1,6 @@
 struct Solution{
     int numRabbits(vector<int>& answers) {
-        int16_t v[1001];
+        int8_t v[1001];
         int res=0;
         for(int i:answers) if(!i || ++v[i]%(i+1)==1) res+=(i+1);
         return res;
